type(voiceSearch): add types to VoiceSearchHelper
